feat(Movie): Movie를 받아오는 Repository와 안드로이드 리소스를 추가해주는 매퍼 생성
- Movie에 대한 정보를 data 모듈로부터 받아오는 MovieRepository 생성. MovieEntity를 Entity로 변경

- Movie에 thumbnail과 poster가 없으므로 Mapper에서 확장 프로퍼티로 넣어주므로써 도메인의 Movie를 app 모듈에서 그대로 사용 가능
